+++ config/config.exs
@@ -476,7 +476,7 @@ config :screens,
 
 config :screens, :screens_by_alert,
   cache_module: Screens.ScreensByAlert.GenServer,
-  screen_data_fn: &Screens.V2.ScreenData.by_screen_id/2,
+  screen_data_fn: &Screens.V2.ScreenData.get/2,
   screens_by_alert_ttl_seconds: 40,
   screens_last_updated_ttl_seconds: 3600,
   screens_ttl_seconds: 40
