+++ config/test.exs
@@ -156,7 +156,7 @@ config :logger, level: :warning
 
 config :screens, :screens_by_alert,
   cache_module: Screens.ScreensByAlert.GenServer,
-  screen_data_fn: &Screens.V2.MockScreenData.by_screen_id/2,
+  screen_data_fn: &Screens.V2.MockScreenData.get/2,
   screens_by_alert_ttl_seconds: 2,
   screens_last_updated_ttl_seconds: 2,
   screens_ttl_seconds: 1
