+++ lib/screens/v2/candidate_generator.ex
@@ -1,7 +1,8 @@
 defmodule Screens.V2.CandidateGenerator do
   @moduledoc false
 
-  alias Screens.V2.ScreenData
+  alias Screens.V2.WidgetInstance
+  alias ScreensConfig.Screen
 
   @doc """
   Returns the template for this screen type.
@@ -12,16 +13,13 @@ defmodule Screens.V2.CandidateGenerator do
   Fetches data and returns a list of candidate widget instances to be
   considered for placement on the template.
   """
-  @callback candidate_instances(ScreenData.config(), keyword()) ::
-              ScreenData.candidate_instances()
+  @callback candidate_instances(Screen.t(), keyword()) :: [WidgetInstance.t()]
 
   @doc """
   Receives the finalized list of widget instances that were placed on
   the template and have defined audio equivalence, as well as screen config,
   and returns a list of zero or more audio-only widgets to be added to the readout.
   """
-  @callback audio_only_instances(
-              widgets :: ScreenData.candidate_instances(),
-              config :: ScreenData.config()
-            ) :: ScreenData.candidate_instances()
+  @callback audio_only_instances(widgets :: [WidgetInstance.t()], config :: Screen.t()) ::
+              [WidgetInstance.t()]
 end
