+++ lib/screens/v2/screen_data.ex
@@ -1,429 +1,66 @@
 defmodule Screens.V2.ScreenData do
   @moduledoc false
 
-  require Logger
-
   alias Screens.Config.Cache
   alias Screens.ScreensByAlert
-  alias Screens.Util
   alias Screens.V2.AlertsWidget
-  alias Screens.V2.ScreenData.Parameters
+  alias Screens.V2.ScreenData.{Layout, Parameters}
   alias Screens.V2.Template
   alias Screens.V2.WidgetInstance
+  alias ScreensConfig.Screen
 
-  import Screens.V2.Template.Guards
+  import Screens.V2.Template.Guards, only: [is_slot_id: 1, is_paged_slot_id: 1]
 
+  @type t :: %{type: atom()}
+  @type simulation_data :: %{full_page: t(), flex_zone: [t()]}
   @type screen_id :: String.t()
-  @type config :: ScreensConfig.Screen.t()
-  @type candidate_instances :: list(WidgetInstance.t())
-  @type selected_instances_map :: %{Template.slot_id() => WidgetInstance.t()}
-  @type non_paged_selected_instances_map :: %{Template.non_paged_slot_id() => WidgetInstance.t()}
-  @type serializable_map :: %{type: atom()}
-  @type response_map :: %{
-          data: serializable_map() | nil,
-          force_reload: boolean(),
-          disabled: boolean(),
-          # Mercury uses this to help optimize data refreshes
-          # https://app.asana.com/0/1185117109217413/1205234924224431/f
-          last_deploy_timestamp: DateTime.t()
-        }
-  @type paging_metadata :: %{
-          Template.non_paged_slot_id() =>
-            {page_index :: non_neg_integer(), num_pages :: pos_integer()}
-        }
+  @type options :: [
+          logging_options: %{atom() => term()},
+          pending_config: Screen.t(),
+          update_visible_alerts?: boolean()
+        ]
 
-  @spec outdated_response() :: response_map()
-  def outdated_response, do: response(force_reload: true)
+  @spec get(screen_id()) :: t()
+  @spec get(screen_id(), options()) :: t()
+  def get(screen_id, opts \\ []) do
+    config = get_config(screen_id, opts)
 
-  @spec disabled_response() :: response_map()
-  def disabled_response, do: response(disabled: true)
-
-  @spec by_screen_id(screen_id(), Keyword.t()) :: response_map() | nil
-  def by_screen_id(screen_id, opts \\ []) do
-    config = get_config(screen_id)
-    refresh_rate = Parameters.get_refresh_rate(config)
-
-    layout_and_widgets =
-      config
-      |> fetch_data(opts)
-      |> tap(&cache_visible_alert_widgets(&1, screen_id, config.hidden_from_screenplay))
-      |> resolve_paging(refresh_rate)
-
-    unless opts[:skip_serialize] do
-      data = serialize(layout_and_widgets)
-      last_deploy_timestamp = Cache.last_deploy_timestamp()
-      response(data: data, last_deploy_timestamp: last_deploy_timestamp)
-    end
+    screen_id
+    |> generate_layout(config, opts)
+    |> resolve_paging(config)
+    |> serialize()
   end
 
-  @spec simulation_data_by_screen_id(screen_id()) :: response_map()
-  def simulation_data_by_screen_id(screen_id) do
-    config = get_config(screen_id)
-    refresh_rate = Parameters.get_refresh_rate(config)
-    screen_data = fetch_data(config)
+  @spec simulation(screen_id()) :: simulation_data()
+  @spec simulation(screen_id(), options()) :: simulation_data()
+  def simulation(screen_id, opts \\ []) do
+    config = get_config(screen_id, opts)
+    layout = generate_layout(screen_id, config, opts)
 
-    full_page_data = screen_data |> resolve_paging(refresh_rate) |> serialize()
-    paged_slot_data = screen_data |> get_paged_slots() |> serialize_paged_slots(config.app_id)
-
-    response(data: %{full_page: full_page_data, flex_zone: paged_slot_data})
+    %{
+      full_page: layout |> resolve_paging(config) |> serialize(),
+      flex_zone: layout |> serialize_paged_slots(config.app_id)
+    }
   end
 
-  @spec pending_data_by_screen_config(ScreensConfig.Screen.t()) :: response_map()
-  def pending_data_by_screen_config(pending_screen_config, opts \\ []) do
-    refresh_rate = Parameters.get_refresh_rate(pending_screen_config)
+  defp get_config(screen_id, opts),
+    do: Keyword.get_lazy(opts, :pending_config, fn -> Cache.screen(screen_id) end)
 
-    layout_and_widgets =
-      pending_screen_config
-      |> fetch_data(opts)
-      |> resolve_paging(refresh_rate)
-
-    data = serialize(layout_and_widgets)
-
-    response(data: data)
+  defp generate_layout(screen_id, config, opts) do
+    config
+    |> Layout.generate(opts)
+    |> tap(
+      &if(
+        Keyword.get(opts, :update_visible_alerts?, false),
+        do: update_visible_alerts(&1, screen_id, config)
+      )
+    )
   end
 
-  @spec pending_simulation_data_by_screen_config(ScreensConfig.Screen.t()) :: response_map()
-  def pending_simulation_data_by_screen_config(pending_screen_config) do
-    refresh_rate = Parameters.get_refresh_rate(pending_screen_config)
-    screen_data = fetch_data(pending_screen_config)
+  defp resolve_paging(layout, config),
+    do: Layout.resolve_paging(layout, Parameters.get_refresh_rate(config))
 
-    full_page_data = screen_data |> resolve_paging(refresh_rate) |> serialize()
-
-    paged_slot_data =
-      screen_data |> get_paged_slots() |> serialize_paged_slots(pending_screen_config.app_id)
-
-    response(data: %{full_page: full_page_data, flex_zone: paged_slot_data})
-  end
-
-  @spec fetch_data(ScreensConfig.Screen.t()) :: {Template.layout(), selected_instances_map()}
-  def fetch_data(config, opts \\ []) do
-    candidate_generator = Parameters.get_candidate_generator(config)
-    screen_template = candidate_generator.screen_template()
-
-    candidate_instances =
-      config
-      |> candidate_generator.candidate_instances(opts)
-      |> Enum.filter(&WidgetInstance.valid_candidate?/1)
-
-    pick_instances(screen_template, candidate_instances)
-  end
-
-  @spec get_config(screen_id()) :: config()
-  def get_config(screen_id) do
-    Screens.Config.Cache.screen(screen_id)
-  end
-
-  @spec pick_instances(Template.template(), candidate_instances()) ::
-          {Template.layout(), selected_instances_map()}
-  def pick_instances(screen_template, candidate_instances) do
-    prioritized_instances = Enum.sort_by(candidate_instances, &WidgetInstance.priority/1)
-
-    # N.B. Each template can place each instance it contains in a different place, so we need to
-    # store a mapping from slot_id to instance for each template.
-    candidate_placements =
-      screen_template
-      |> Template.slot_combinations()
-      |> Enum.map(fn t -> {t, %{}} end)
-      |> Enum.into(%{})
-
-    {{_, {slot_id, {layout_type, _children}} = selected_layout}, selected_instances} =
-      prioritized_instances
-      |> Enum.reduce(candidate_placements, &place_instance/2)
-      |> log_mismatched_placement_widgets()
-      |> select_best_placement()
-
-    filtered_children =
-      selected_layout
-      |> filter_empty_slots(Map.keys(selected_instances))
-
-    {{slot_id, {layout_type, filtered_children}}, selected_instances}
-  end
-
-  defp place_instance(instance, placements) do
-    instance_slots = WidgetInstance.slot_names(instance)
-    live_templates = Map.keys(placements)
-    placeable_templates = get_valid_templates(live_templates, instance_slots)
-
-    case placeable_templates do
-      [] ->
-        # If none of the remaining templates can hold the current instance, don't include that
-        # instance, and continue with the same live templates and placements.
-        placements
-
-      _ ->
-        # If at least one of the remaining templates can hold the current instance, throw out all
-        # templates which can't. For those which can, choose a valid slot and place the instance
-        # there, adding it to the slot_id => instance mapping for that template, and removing it
-        # from the list of unoccupied slot_ids.
-        placeable_templates
-        |> Enum.map(fn t ->
-          chosen_slot = get_first_slot(t, instance_slots)
-
-          updated_placement =
-            placements
-            |> Map.get(t)
-            |> Map.put(chosen_slot, instance)
-
-          {slots, layout} = t
-          new_t = {slots -- [chosen_slot], layout}
-
-          {new_t, updated_placement}
-        end)
-        |> Enum.into(%{})
-    end
-  end
-
-  defp get_valid_templates(templates, instance_slots) do
-    Enum.filter(templates, &template_is_placeable?(&1, instance_slots))
-  end
-
-  defp get_first_slot(template, instance_slots) do
-    template
-    |> get_valid_slots(instance_slots)
-    |> hd()
-  end
-
-  defp template_is_placeable?(template, instance_slots) do
-    matching_slots = get_valid_slots(template, instance_slots)
-    length(matching_slots) > 0
-  end
-
-  defp get_valid_slots(template, instance_slots) do
-    {template_slots, _} = template
-
-    # N.B. The slots are sorted so that paged regions have their earlier pages filled first.
-    # e.g. [{0, :paged_region1}, {0, :paged_region2}, {1, :paged_region1}, {1, :paged_region2}]
-    sorted_slot_list_intersection(template_slots, instance_slots)
-  end
-
-  defp log_mismatched_placement_widgets(placements) when map_size(placements) < 2, do: placements
-
-  defp log_mismatched_placement_widgets(placements) do
-    [first_widget_set | widget_sets] =
-      placements
-      |> Map.values()
-      |> Enum.map(&Map.values/1)
-      |> Enum.map(&MapSet.new/1)
-
-    _ =
-      if Enum.any?(widget_sets, &(not MapSet.equal?(first_widget_set, &1))) do
-        Logger.info("[mismatched widget placements]")
-      end
-
-    placements
-  end
-
-  defp select_best_placement(placements) when map_size(placements) < 2 do
-    placements
-    |> Map.to_list()
-    |> hd()
-  end
-
-  defp select_best_placement(placements) do
-    # When multiple valid placements are produced due to paging, prefer the one
-    # that has higher-priority instances placed on earlier pages.
-    # Each placement is compared by sorting the mapping by instance priority,
-    # then mapping each entry to its slot's page index (or nil if not paged).
-    # Example min_by comparison key: [nil, 0, 1, 0, 0]
-    placements
-    |> Enum.min_by(fn {_, slot_to_instance} ->
-      slot_to_instance
-      |> Enum.sort_by(fn {_, instance} -> WidgetInstance.priority(instance) end)
-      |> Enum.map(fn
-        {{page_index, _slot}, _} -> page_index
-        {_slot, _} -> nil
-      end)
-    end)
-  end
-
-  @spec sorted_slot_list_intersection(
-          list(Template.slot_id()),
-          list(Template.non_paged_slot_id())
-        ) ::
-          list(Template.slot_id())
-  def sorted_slot_list_intersection(template_slots, instance_slots) do
-    for t_slot <- template_slots,
-        i_slot <- instance_slots,
-        Template.slots_match?(t_slot, i_slot) do
-      t_slot
-    end
-    |> Enum.sort(&Template.slot_precedes_or_equal?/2)
-  end
-
-  @spec resolve_paging(
-          {Template.layout(), selected_instances_map()},
-          integer() | nil,
-          DateTime.t()
-        ) ::
-          {Template.non_paged_layout(), non_paged_selected_instances_map(), paging_metadata()}
-  def resolve_paging(layout_and_instances, refresh_rate, now \\ DateTime.utc_now())
-
-  def resolve_paging(layout_and_instances, nil, _now) do
-    Tuple.append(layout_and_instances, %{})
-  end
-
-  def resolve_paging({layout, instance_map}, refresh_rate, now) do
-    {unpaged_layout, selected_paged_slot_ids, paging_metadata} =
-      choose_visible_slot_ids(layout, refresh_rate, now)
-
-    # Now filter instance map by set of paged slot ids, then unpage
-    instance_map =
-      instance_map
-      |> Enum.filter(fn
-        {slot_id, _instance} when is_paged_slot_id(slot_id) ->
-          slot_id in selected_paged_slot_ids
-
-        _ ->
-          true
-      end)
-      |> Enum.map(fn {slot_id, instance} -> {Template.unpage(slot_id), instance} end)
-      |> Enum.into(%{})
-
-    {unpaged_layout, instance_map, paging_metadata}
-  end
-
-  def get_paged_slots({layout, instance_map}) do
-    paged_instances =
-      instance_map
-      |> Map.filter(fn
-        {slot_id, _instance} when is_paged_slot_id(slot_id) ->
-          true
-
-        _ ->
-          false
-      end)
-
-    {paged_instances, layout}
-  end
-
-  defp select_page_index(num_pages, refresh_rate, now) do
-    seconds_since_midnight = now.hour * 60 * 60 + now.minute * 60 + now.second
-    periods_since_midnight = div(seconds_since_midnight, refresh_rate)
-    rem(periods_since_midnight, num_pages)
-  end
-
-  # Function to remove slots from the layout if there are no candidates available to populate it
-  # Specifically added to help introduce variable paging
-  defp filter_empty_slots({_slot_id, {_layout_type, children}}, selected_instances_slots) do
-    children
-    |> Enum.map(fn
-      # Static slots: :main_content, :header, etc.
-      slot_id when is_atom(slot_id) ->
-        if slot_id in selected_instances_slots do
-          slot_id
-        end
-
-      # Paged slots: :flex_zone
-      nested_child when is_paged_slot_id(nested_child) ->
-        if nested_child in selected_instances_slots do
-          nested_child
-        end
-
-      # A nested layout. Take the nested layout and feed it back into this function to process its children.
-      {slot_id, {layout_type, _children}} = nested_layout ->
-        case filter_empty_slots(nested_layout, selected_instances_slots) do
-          [nil] ->
-            nil
-
-          child ->
-            {slot_id, {layout_type, child}}
-        end
-    end)
-    # Remove all nil/empty pages from the original children.
-    |> Enum.filter(fn
-      nil -> false
-      {_, {_, []}} -> false
-      _ -> true
-    end)
-  end
-
-  @spec choose_visible_slot_ids(Template.layout(), integer(), DateTime.t()) ::
-          {Template.non_paged_layout(), MapSet.t(Template.paged_slot_id()), paging_metadata()}
-  defp choose_visible_slot_ids(layout, refresh_rate, now)
-
-  defp choose_visible_slot_ids(slot_id, _refresh_rate, _now) when is_non_paged_slot_id(slot_id) do
-    {slot_id, MapSet.new(), %{}}
-  end
-
-  defp choose_visible_slot_ids({slot_id, {layout_type, children}}, refresh_rate, now) do
-    {children, selected_paged_slot_ids, paging_metadata} =
-      choose_visible_slot_ids(children, refresh_rate, now)
-
-    {{slot_id, {layout_type, children}}, selected_paged_slot_ids, paging_metadata}
-  end
-
-  defp choose_visible_slot_ids(layouts, refresh_rate, now) when is_list(layouts) do
-    {selected_paged_slot_ids, paging_metadata_entries} =
-      layouts
-      |> Enum.filter(fn
-        layout when is_paged(layout) -> true
-        _ -> false
-      end)
-      |> Enum.group_by(&Template.unpage/1, &Template.get_page/1)
-      |> Enum.map(fn {slot_id, page_indexes} -> {slot_id, Enum.max(page_indexes) + 1} end)
-      |> Enum.map(fn {slot_id, num_pages} ->
-        selected_page_index = select_page_index(num_pages, refresh_rate, now)
-        selected_paged_slot_id = {selected_page_index, slot_id}
-        paging_metadata_entry = {slot_id, {selected_page_index, num_pages}}
-
-        {selected_paged_slot_id, paging_metadata_entry}
-      end)
-      |> Enum.unzip()
-
-    paging_metadata = Map.new(paging_metadata_entries)
-
-    selected_layouts =
-      Enum.filter(layouts, fn
-        layout when is_paged(layout) -> Template.get_slot_id(layout) in selected_paged_slot_ids
-        _ -> true
-      end)
-
-    # Now we have the list of layouts to keep, but still need to unpage them
-    # and create a set of the paged slot ids to keep in the instance map.
-    # We also need to recurse on non-paged layouts in case they contain paging deeper down.
-    {unpaged_layouts, paged_slot_sets, paging_metadata_maps} =
-      selected_layouts
-      |> Enum.map(fn
-        layout when is_paged(layout) ->
-          layout
-          |> unpage_layout_and_track_slots()
-          |> Tuple.append(%{})
-
-        layout ->
-          choose_visible_slot_ids(layout, refresh_rate, now)
-      end)
-      |> Util.unzip3()
-
-    paged_slot_set = Enum.reduce(paged_slot_sets, &MapSet.union/2)
-
-    paging_metadata = Enum.reduce(paging_metadata_maps, paging_metadata, &Map.merge/2)
-
-    {unpaged_layouts, paged_slot_set, paging_metadata}
-  end
-
-  defp unpage_layout_and_track_slots(slot_id) when is_paged_slot_id(slot_id) do
-    {Template.unpage(slot_id), MapSet.new([slot_id])}
-  end
-
-  defp unpage_layout_and_track_slots({slot_id, {layout_type, children}})
-       when is_paged_slot_id(slot_id) do
-    {children, paged_slot_sets} =
-      children
-      |> Enum.map(&unpage_layout_and_track_slots/1)
-      |> Enum.unzip()
-
-    paged_slot_set = Enum.reduce(paged_slot_sets, &MapSet.union/2)
-
-    {{Template.unpage(slot_id), {layout_type, children}}, paged_slot_set}
-  end
-
-  defp unpage_layout_and_track_slots(non_paged_layout) do
-    {non_paged_layout, MapSet.new()}
-  end
-
-  @spec serialize(
-          {Template.non_paged_layout(), non_paged_selected_instances_map(), paging_metadata()}
-        ) :: map() | nil
+  @spec serialize(Layout.non_paged()) :: map() | nil
   def serialize({layout, instance_map, paging_metadata}) do
     serialized_instance_map =
       instance_map
@@ -433,14 +70,12 @@ defmodule Screens.V2.ScreenData do
     Template.position_widget_instances(layout, serialized_instance_map, paging_metadata)
   end
 
-  defp serialize_paged_slots({instance_map, layout}, app_id) do
-    # instance_map looks like:
-    # %{{page_index, slot_id} => instance}
-
-    # we want:
-    # %{page_index => %{slot_id => instance}}
-
+  defp serialize_paged_slots({layout, instance_map}, app_id) do
     instance_map
+    |> Map.filter(fn
+      {slot_id, _instance} when is_paged_slot_id(slot_id) -> true
+      _ -> false
+    end)
     |> Enum.group_by(
       &paged_slot_key(&1, app_id),
       fn {paged_slot_id, instance} -> {Template.unpage(paged_slot_id), instance} end
@@ -472,16 +107,6 @@ defmodule Screens.V2.ScreenData do
     |> Map.merge(%{type: WidgetInstance.widget_type(instance)})
   end
 
-  @spec response(keyword()) :: response_map()
-  defp response(fields) do
-    %{
-      data: Keyword.get(fields, :data, nil),
-      force_reload: Keyword.get(fields, :force_reload, false),
-      disabled: Keyword.get(fields, :disabled, false),
-      last_deploy_timestamp: Keyword.get(fields, :last_deploy_timestamp, nil)
-    }
-  end
-
   @spec get_containing_slot(Template.layout(), list(Template.non_paged_slot_id())) ::
           Template.non_paged_slot_id()
 
@@ -511,9 +136,9 @@ defmodule Screens.V2.ScreenData do
     end
   end
 
-  def cache_visible_alert_widgets(_, _, true), do: nil
+  def update_visible_alerts(_, _, %Screen{hidden_from_screenplay: true}), do: :ok
 
-  def cache_visible_alert_widgets({_layout, instance_map}, screen_id, _) do
+  def update_visible_alerts({_layout, instance_map}, screen_id, _config) do
     alert_ids =
       instance_map
       |> Map.values()
